fix: Don't delete ClusterServingRuntime in namespace-scope mode (#421)
Set NAMESPACE_SCOPE environment to true or false in install script
and test NAMESPACE_SCOPE environment variable in delete script
to check whether ClusterServingRuntime should be deleted.

Resolves #420

---------

+++ scripts/delete.sh
@@ -131,9 +131,15 @@ if [[ "$crb_ns" == "$namespace" ]]; then
   echo "deleting cluster scope RBAC"
   kustomize build rbac/cluster-scope | kubectl delete -f - --ignore-not-found=true
 fi
+
+# Determine whether deployment is namespace-scoped before deleting runtime resources
+is_namespace_scoped=$(kubectl exec deploy/modelmesh-controller -- printenv NAMESPACE_SCOPE 2> /dev/null || echo "false") || :
 kustomize build default | kubectl delete -f - --ignore-not-found=true
 kustomize build rbac/namespace-scope | kubectl delete -f - --ignore-not-found=true
-kustomize build runtimes ${kustomize_load_restrictor_arg} | kubectl delete -f - --ignore-not-found=true
+if [[ ! "$is_namespace_scoped" == "true" ]]; then
+  kustomize build runtimes ${kustomize_load_restrictor_arg} | kubectl delete -f - --ignore-not-found=true
+fi
+
 kubectl delete -f dependencies/quickstart.yaml --ignore-not-found=true
 kubectl delete -f dependencies/fvt.yaml --ignore-not-found=true
